refactor: 💡 put lecture exceptions in a new package
✅ Closes: #52
